services/hello-world/requirements.txt
grpcio-tools==1.29.0
